Added start menu
changed clock class to support negativ start values

+++ Game/Game/Clock.h
@@ -3,7 +3,7 @@
 class Clock
 {
 private:
-	Uint32 m_startPoint_;
+	int64_t m_startPoint_;
 	unsigned int m_clockInterval_;
 	bool m_setStartPointAuto_;		//Default: true, resets the start point when check returns true
 public:
@@ -11,8 +11,8 @@ public:
 	Clock(unsigned int m_clockInterval_, bool m_setStartPointAuto_);
 	Clock();
 	bool checkClockState();
-	inline void setStartPoint(Uint32 m_startPoint_) { this->m_startPoint_ = m_startPoint_; }
-	inline Uint32 getStartPoint() { return m_startPoint_; }
+	inline void setStartPoint(int64_t m_startPoint_) { this->m_startPoint_ = m_startPoint_; }
+	inline int64_t getStartPoint() { return m_startPoint_; }
 	inline void setInterval(unsigned int m_clockInterval_) { this->m_clockInterval_ = m_clockInterval_; }
 };
